refactor(Q1): clippy --fix mechanical sweep — redundant_field_names
cargo clippy --fix on the workspace under
--features hip,hip_serve,hip_sweep, accepting only redundant-field-
names rewrites (the only autofix that ran cleanly across the diff).
36 sites: `MyStruct { x: x, y: y }` → `MyStruct { x, y }`.

Clippy delta:
- total warnings: 320 → 277 (−43)
- too_many_arguments: unchanged
- errors: 0

Tests: backend-hip 21/21, ops 2/2, model-ops 30/30, quant 29/29,
runtime 48/48, server 86/86.

+++ crates/ops/tests/moe_sort_smoke.rs
@@ -299,7 +299,7 @@ fn moe_sort_by_expert_padded_groups_in_multiples_of_8() -> Result<()> {
     // Verify padded_offsets prefix-sum = sum of ceil(counts[e]/8)*8.
     let mut cpu_padded = vec![0i32; n_experts + 1];
     for e in 0..n_experts {
-        let padded_c = ((counts[e] + 7) & !7) as i32;
+        let padded_c = ((counts[e] + 7) & !7);
         cpu_padded[e + 1] = cpu_padded[e] + padded_c;
     }
     assert_eq!(padded_off, cpu_padded, "padded_offsets mismatch");
